+++ packages/client/src/index.ts
@@ -1,0 +1,6 @@
+export * from './client'
+export * from './enums'
+export * from './errors'
+export * from './types'
+export * from './utils'
+export * from './wallet'